requirements.txt
backcall==0.2.0
black==21.11b1
click==8.0.3
dataclasses==0.8
decorator==5.1.0
Django==2.2.12
django-cors-headers==3.10.0
importlib-metadata==4.8.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.18.5
opencv-python==4.4.0.42
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pydicom==2.0.0
Pygments==2.10.0
pytz==2021.3
regex==2021.11.10
scipy==1.2.1
SimpleITK==2.0.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
traitlets==4.3.3
typed-ast==1.5.0
typing_extensions==4.0.0
uWSGI==2.0.20
wcwidth==0.2.5
zipp==3.6.0
